Correct guessing
Updated the guess handler to get value from the guess input box. Server
was erroring on guess.toLowerCase(), in 'guess' handler, so commented
those out. PlayCtrl got a handler for 'correctGuess'. Currently that
handler doesn't do anything, because I don't know how to modify a value
in an HTML table.